Code utilized in the project.
First, the IMF is defined and, since we are utilizing the Monte Carlo numerical method, it has applied to it statistical weights according to the mass. AS it is necessary, a constant SFR is made up, and then for each mass an age is assigned from this SFR that covers the age of the Milky Way. Later, utilizing the mass-luminosity relation for the MS taken from the guide, we can calculate the lifetime in the MS for each star and, since it depends on its mass, some of them can be catalogued as remnants, and to those, we take the equations from Kalirai et al. (2008) and Raithel et al. (2018) to classify them in the three main groups of remnants: White Dwarfs, Neutron Stars and Black Holes. Finally, a series of graphics and histograms are made to see the proportions and distributions of the masses and ages for these groups.